Added display of co2 ppm values: Updated output functions for CO2 visualization and added the CO2 measurement to the lower left corner selection field. The sensor provides the value as factor 10 of ppm that's why the data type had to be changed to 32bit if ppm should be available without scaling every time.
Cleanup sensor dialog:
The HUD battery was displayed by default and some sensor combinations were not displayed correctly. The refresh function was updated to fix these issues.

+++ Common/Inc/data_central.h
@@ -83,7 +83,7 @@ typedef struct
 
 typedef struct
 {
-		uint16_t CO2_ppm;
+		uint32_t CO2_ppm;
 		uint16_t signalStrength;
 } 	SCO2Sensor;
